Embed zip in nupkg instead of downloading from GitHub

+++ chocolatey/packAndLocalInstall.ps1
@@ -6,15 +6,20 @@ $version = $nuspec.package.metadata.version
 $tag = "v$version"
 
 if ($Force) {
-    git tag -f $tag
-    git push -f $Remote $tag
+    git tag --force --annotate --sign $tag -m "Release $tag"
+    git push --force $Remote $tag
 }
 elseif (!$(git ls-remote $Remote $tag)) {
     Write-Warning "'$Remote/$tag' not found! Use -Force to create tag at HEAD."
     return
 }
 
+Invoke-WebRequest -Uri "https://github.com/tygerbytes/posh-dynargs/zipball/v$version" -UseBasicParsing -OutFile "posh-dynargs-v$version.zip"
+
 choco pack poshdynargs.nuspec
-choco install -f -y poshdynargs -pre --version=$version -s "'.;chocolatey'"
+
+if ((Read-Host "Install the package? (y/n)") -eq "y") {
+    choco install -f -y poshdynargs -pre --version=$version -s "'.;chocolatey'"
+}
 
 Pop-Location
